Add soft k-means and dynamic n_out to routing algorithm.
This update adds a new optional argument to the routing module, p_model,
for specifying how to to compute the probability of input capsule votes
at each output capsule.  Two values for p_model are now supported:

  ‘gaussian’ for a Gaussian mixture model, the default, and
  ‘skm’ for a soft k-means model, a new option.

See the model docstring for usage instructions.

This update also makes it possible to specify the number of output
capsules dynamically, if it is not specified in advanced as an argument,
enabling applications in which the number of output capsules can be
programmatically changed on-the-fly. Example:

  >>> a_inp = torch.randn(100)  # 100 input scores
  >>> mu_inp = torch.randn(100, 4, 4)  # 100 capsules of shape 4 x 4
  >>> m = Routing(d_cov=4, d_inp=4, d_out=4, n_inp=100)  # variable n_out
  >>> a_out, mu_out, sig2_out = m(a_inp, mu_inp, n_out=42)
  >>> print(mu_out)  # 42 capsules of shape 4 x 4

If n_out is not an input to the forward pass, the number of output
capsules is set equal to the number of input capsules, as before.