Use a parameter to break symmetry when n_out is variable.
This parametrized approach to breaking symmetry when n_out is variable
is a bit more elegant and flexible, and works well in our testing.

NOTE: We expect that most applications will *never* use this. They will
use fixed n_out instead, like our sample applications to smallNORB and
SST. In our paper, we mention the use of variable n_out only briefly as
a possible extension to the routing algorithm. Nonetheless, in the
current draft we neglect to point out that when we apply the same linear
transformation to all input capsules, we must break symmetry via other
means -- for example, by adding a different bias to each output capsule,
which is what we are doing here. This is an important point, so we will
add a sentence about it to the paper when we next udpate it on arXiv, as
we continue to receive and incorporate feedback from other researchers.