Avoid relying on PyTorch einsum's implicit order of contraction.
With this commit, we now explicitly contract `phi` and `scaled_x_inp` over
index i before contracting them with other tensors over index d to compute
updated output vector states `x_out` at the end of each iteration in
EfficientVectorRouting. Before this commit, we were relying on PyTorch
einsum's conventional order of contraction. The computation remains the
same, but the order of contraction over indices i and d is now explicit.

As an added benefit, the computation of `x_out` in EfficientVectorRouting
is now easier to compare to the expression in line 18 of Algorithm 2 in the
2022 paper: Each einsum operator corresponds to one summation symbol in the
expression. See Appendix C in the same paper for the derivation.